Implement mutation program
mutate.cpp is a naive mutation program that takes a plant genome and
randomly changes hex characters in the genome string. If this results in
an invalid genome, the program discards the mutant genome and tries
again. Comments and whitespace are preserved and ignored. mutate.cpp
takes two arguments, a path to an existing genome file and a path to
save the mutant genome at.