Implement cross program and makefile
cross.cpp crosses two parent genomes of equal gene length and produces a
child genome of the same length by randomly choosing a parent to inherit
each gene from. The program takes three arguments, a file containing the
genome string of parent one, a file containing the genome of the second
parent, and the file location to save the genome string of the child. The
output genome string will be unformatted.

This commit includes a makefile to make compiling render, cross, and
mutate easier and faster.

The flowergenomes directory has also been removed in this commit. Seems
like this would cut down on mess in the repository and if anyone else
ever actually uses this code, they could just write their own.