Implement web interface and refactor C++ programs
The web interface is implement with a simple NodeJS HTTP server that
creates child processes on the server to run render, mutate, and cross.
Input for these programs are piped from the request object and the ouput
is piped into the response object.

In order to facillitate piping, render.cpp, mutate.cpp, and cross.cpp
have been refactored to take input from stdin and print output to stdout.
Additionally, when the flag "-w" is passed to render.cpp it outputs svg
in an HTML friendly format. When "-f" is passed to mutate.cpp, it
overrides the check for a valid genome output. When "-s" is passed to
cross.cpp, it takes input from stdin in the form of two genome strings
concatenated with " $ ". Otherwise, it can still take filename arguments.

A bug involving jumping to negative gene numbers has been fixed.

In terms of housecleaning, all C++ source code has been moved to the src
directory and a bin directory has been created to hold the Makefile,
object files, and binaries.