Redesign web interface and fix bugs
The front-end interface written in NodeJS has been redesigned to look
nicer and function better. Users can now tab between three different
genomes and their outputs. The render and mutate buttons operate on the
currently selected genome and the cross button crosses the two genomes on
either side of the current genome tab. An alert pop-up reports any errors
from the server.

A couple bugs mutate.cpp and render.cpp that caused the programs to hang
indefinitely have been fixed. Additionally a strange bug in index.js was
resolved. Apparently, for certain genomes that outputted many lines to
stderr, not piping the child process' stderr to /dev/null causes
render.cpp to hang. I don't know if this was because the stderr buffer
was somehow filled up and needed to be dumped, or for some other reason,
but bit bucketing stderr seems to have solved the issue.