Refactor error handling and fix spawn cell bug
Error handling in the C++ programs has been refactored to use data
members in order to detect errors in Genome construction. Previously,
the Genome.cpp constructor threw a runtime exception if the input was
malformed. Now it simply sets m_err and returns early. Plant.cpp now
checks m_err of its Genome object and sets its own m_err. If an error
occurs in render.cpp, mutate.cpp, or cross.cpp, it is reported to stdout
with an "ERROR: " prefix.

This change was primarily made due to the difficulty with communicating
error details to the client. Because the output of each child process is
piped directly to the response object, the only convenient way to send
errors is through stdout. Client-side logic checks the response from the
server for these errors.

Additionally, a bug in the spawn cell instruction that allowed cells to
be spawned with starting genes greater than the number of genes in the
genome. The instruction now modulos the argument with the total number of
genes beforing spawning the new cell.