Polish up front end and implement throttling
A new user help panel has been added to the webpage, which provides
documentation as well as several example genomes.

On the front end, a small change has been made so that the user cannot
create multiple requests for render.cpp when the genome has not been
modified. On the back end, child processes are now created one at a time.
The server queues requests and waits for the current child to exit before
creating a new child to fulfill the request. Currently this change
specifically accomodates single core hardware. It is to be seen what kind
of performance impact this will have.